move blending and gradient to their own modules
