+++ node-graph/gcore/src/lib.rs
@@ -1,38 +1,36 @@
 #[macro_use]
 extern crate log;
 
-pub use crate as graphene_core;
-pub use ctor;
-pub use num_traits;
-
 pub mod animation;
+pub mod blending;
 pub mod consts;
 pub mod context;
 pub mod generic;
+pub mod gradient;
+mod graphic_element;
 pub mod instances;
 pub mod logic;
+pub mod memo;
 pub mod misc;
 pub mod ops;
+pub mod raster;
 pub mod raster_types;
+pub mod registry;
 pub mod structural;
 pub mod text;
+pub mod transform;
 pub mod uuid;
 pub mod value;
-
-pub mod memo;
-
-pub mod raster;
-pub mod transform;
-
-mod graphic_element;
-pub use graphic_element::*;
 pub mod vector;
 
-pub mod registry;
-
+pub use crate as graphene_core;
+pub use blending::*;
 pub use context::*;
+pub use ctor;
 pub use dyn_any::{StaticTypeSized, WasmNotSend, WasmNotSync};
+pub use graphic_element::*;
 pub use memo::MemoHash;
+pub use num_traits;
 pub use raster::Color;
 use std::any::TypeId;
 use std::future::Future;
